Providing reflectance scaling
The L1b data needs to be converted from sensor counts to TOA
reflectance. The L1 product does have some scaling information,
but this information is lost when gridding it. I have now changed
my gridding script to store this information in the geotiffs, and
if that is the case, the class now retrieves the reflectance scales
and offsets per band (the latter are usually 0).